helpers: added new helper function that creates and returns a media player used to play sound files
Honestly in my opinion not that useful of a function.

Eitherway to use the function just create a private variable called mediaplayer in whichever activity you want to play a sound in, call the helper to get the media player,
attach an onclick listener somewhere that starts the media player. On activity exit stop the mediaplayer prior to exit